feat: migrate aidea to rocco
