+++ apps/api/package.json
@@ -25,7 +25,7 @@
     "@fastify/helmet": "^11.0.0",
     "@fastify/rate-limit": "^8.0.3",
     "@fastify/secure-session": "^6.1.0",
-    "@prisma/client": "^5.0.0",
+    "@prisma/client": "^5.1.1",
     "@segment/analytics-node": "^1.0.0",
     "@sendgrid/mail": "^7.6.1",
     "@sentry/node": "^7.60.1",
@@ -46,7 +46,7 @@
     "eslint-config-ponti": "0.0.0",
     "jest": "29.5.0",
     "nodemon": "^3.0.1",
-    "prisma": "^5.0.0",
+    "prisma": "5.1.1",
     "ts-jest": "29.1.0",
     "ts-node": "10.4.0",
     "tsconfig": "0.0.0",
