+++ apps/web/src/components/App/index.tsx
@@ -56,7 +56,7 @@ function App() {
   return (
     <div id="app" className={`h-100 flex flex-col ${styles.wrap}`}>
       <Header />
-      <main className="flex sm:mt-4 md:mt-8">
+      <main className="flex flex-1 sm:mt-4 md:mt-8">
         <Routes>
           <Route path={AUTHENTICATE_PATH} element={<Authenticate />} />
           {isAuthenticated ? (
