--- packages/ui/Header/AuthNavMenu/components/NavMenuItem/NavMenuItem.module.css
@@ -1,11 +1,0 @@
-.menuItem {
-  margin-right: 10px;
-}
-
-.menuItem:last-child {
-  margin-right: 0;
-}
-
-.menuItem a {
-  text-decoration: none;
-}
